+++ CHANGELOG.md
@@ -117,12 +117,14 @@ It features a lot of new rules from different categories.
 - Adds local Github Action that uses itself for testing
 - Adds official Docker image and docs about it
 
+
 ## 0.12.4
 
 ### Bugfixes
 
 - Fixes bug with `nitpick` colors and new files API
 - Updates `flake8-docstrings`
+
 
 ## 0.12.3
 
@@ -131,6 +133,7 @@ It features a lot of new rules from different categories.
 - Fixes that formatting was failing sometimes when colours were not available
 - Fixes that `1 / number` was not allowed
 - Fixes that `%` operator was allowed for `0` and `1`
+
 
 ## 0.12.2
 
@@ -143,6 +146,7 @@ It features a lot of new rules from different categories.
 ### Bugfixes
 
 - Fixes that `{**a, **b}` was reported as duplicate hash items
+
 
 ## 0.12.1
 
@@ -159,6 +163,7 @@ It features a lot of new rules from different categories.
 
 - Improves `README.md` with `flakehell` and `nitpick` mentions
 - Improves docs all accross the project
+
 
 ## 0.12.0
 
@@ -291,11 +296,13 @@ In this release we had a little focus on:
 - Updates docs about `snake_case` in `Enum` fields
 - Updates docs about `WPS400` and incorrect line number
 
+
 ## 0.11.1
 
 ### Bugfixes
 
 - Now using `pygments` as a direct dependency
+
 
 ## 0.11.0 aka The New Violation Codes
 
@@ -335,6 +342,7 @@ for you to rename your violations with a script.
 - Renames `Incorrect` to `Wrong` where possible
 - Renames `IncorrectlyNestedTernaryViolation` to `NestedTernaryViolation`
 - Renames `IncorectLoopIterTypeViolation` to `WrongLoopIterTypeViolation`
+
 
 ## 0.10.0 aka The Great Compare
 
@@ -393,11 +401,13 @@ big cudos to the developers of this wonderful tool.
 - Adds `Makefile`
 - A lot of minor dependency updates
 
+
 ## 0.9.1
 
 ### Bugfixes
 
 - Fixes issue with `pydocstyle>=4` by glueing its version to `pydocstyle<4`
+
 
 ## 0.9.0
 
@@ -435,11 +445,13 @@ and lots of bug fixes.
 - Better `exclude` rule for `flake8` check
 - Removed warnings from `pytest`
 
+
 ## 0.8.1
 
 ### Bugfixes
 
 - Fixes how `wps_context` is calculated, so `super()` calls are now working
+
 
 ## 0.8.0
 
@@ -471,6 +483,7 @@ and lots of bug fixes.
 - Improves docs for `ProtectedAttributeViolation`
 - Adds docs about `.pyi` files
 
+
 ## 0.7.1
 
 ### Bugfixes
@@ -481,6 +494,7 @@ and lots of bug fixes.
 ### Misc
 
 - Improves docs about using `# type: some` comment in `for` loops
+
 
 ## 0.7.0
 
@@ -545,11 +559,13 @@ and lots of bug fixes.
 - Refactoring: moves `presets` package to the root
 - Improves tests: we now lint our layered architecure with `layer-lint`
 
+
 ## Version 0.6.3
 
 ### Bugfixes
 
 - Fixes an [issue-450](https://github.com/wemake-services/wemake-python-styleguide/issues/450) with `dict`s with just values and no keys
+
 
 ## Version 0.6.2
 
@@ -557,11 +573,13 @@ and lots of bug fixes.
 
 - Fixes a [crash](https://github.com/wemake-services/wemake-python-styleguide/issues/423) with class attributes assignment
 
+
 ## Version 0.6.1
 
 ### Bugfixes
 
 - Fixes a conflict between our plugin and `pyflakes`
+
 
 ## Version 0.6.0
 
@@ -599,12 +617,14 @@ and lots of bug fixes.
 - Improves docs: making contributing section in the `README` more friendly
 - Improves build: changes how CI installs `poetry`
 
+
 ## 0.5.1
 
 ### Bugfixes
 
 - Fixes all possible errors that happen
   because of unset `parent` and `function_type` properties
+
 
 ## 0.5.0
 
@@ -686,6 +706,7 @@ for tests. We also now covering more cases and testing violation texts.
 - Improves tests: now all common patters live in related `conftest.py`
 - Improves docs: now all configuration options are listed in the violations
 
+
 ## 0.3.0 aka The Hacktoberfest Feast
 
 This release was made possible by awesome people who contributed
@@ -754,6 +775,7 @@ to the project during `#hactoberfest`. List of awesome people:
 - Improves types: now we use `final` to indicate what should not be changed
 - Improves types: now we do not have any ugly import hacks
 
+
 ## 0.2.0 aka Revenge of the Async
 
 This release was made possible by awesome people who contributed
@@ -785,6 +807,7 @@ to the project during `#hactoberfest`. List of awesome people:
 - Improves docs: adds naming convention to the `naming.py`
 - Improves docs: multiple typos, bugs, and issues fixes
 - Improves tests: now we are testing `async` comprehensions
+
 
 ## Version 0.1.0
 
@@ -825,6 +848,7 @@ to the project during `#hactoberfest`. List of awesome people:
 - Improves docs: now error pages are split
 - Improves docs: now all `flake8` plugin dependencies are documented
 
+
 ## Version 0.0.15
 
 ### Features
@@ -837,6 +861,8 @@ to the project during `#hactoberfest`. List of awesome people:
 
 - Improves `CONTRIBUTING.md`
 - Moves issues templates to `.github/` folder
+- Update error thrown on `RedundantSubscriptViolation`
+
 
 ## Version 0.0.14
 
@@ -861,6 +887,7 @@ to the project during `#hactoberfest`. List of awesome people:
 - Adds `flake8-print` plugin for development
 - Removes `delegate` concept from the codebase
 
+
 ## Version 0.0.13 aka The Jones Complexity
 
 ### Features
@@ -872,6 +899,7 @@ to the project during `#hactoberfest`. List of awesome people:
 
 - Improves docs: adds detailed installation instructions
 - Removes `flake8-blind-except` plugin
+
 
 ## Version 0.0.12
 
@@ -894,6 +922,7 @@ We have **changed** the error codes for general checks.
 - Improves docs: now docs include `CHANGELOG.md`
 - Improves docs: now we have templates for `bug` and `rule-request`
 - Replaced `pytest-isort` with `flake8-isort`
+
 
 ## Version 0.0.11
 
@@ -936,6 +965,7 @@ There are no new features introduced.
 - Now each error has a link to the corresponding constant (if any)
 - Improved docs with links to the corresponding configuration flags
 
+
 ## Version 0.0.9
 
 This is just a supporting release.
@@ -950,6 +980,7 @@ There are no new features introduced.
 
 - Errors are now tested
 - Complexity tests are refactored
+
 
 ## Version 0.0.8 aka The Complex Complexity
 
@@ -974,6 +1005,7 @@ There are no new features introduced.
 
 - Improved type support for options parsing
 
+
 ## Version 0.0.7
 
 ### Features
@@ -993,6 +1025,7 @@ There are no new features introduced.
 - Tests now cover nested classes' explicit bases
 - Tests now cover nested classes and functions `noqa` comment
 
+
 ## Version 0.0.6
 
 ### Features
@@ -1010,6 +1043,7 @@ There are no new features introduced.
 - Improved typing support
 - Added new documentation sections
 
+
 ## Version 0.0.5
 
 ### Features
@@ -1021,6 +1055,7 @@ There are no new features introduced.
 ### Misc
 
 - We have added a `CONTRIBUTING.md` file to help new contributors
+
 
 ## Version 0.0.4
 
@@ -1036,6 +1071,7 @@ There are no new features introduced.
 - We now have a whitelist for `__future__` imports
 - Imports are now have its own subgroup `Z10`
 - General rules now start from `Z11`
+
 
 ## Version 0.0.3
 
@@ -1056,6 +1092,7 @@ There are no new features introduced.
 - Fixes issue with missing `parent`s :batman:
 - Fixes issue with `_$NAME` patterns being ignored
 
+
 ## Version 0.0.2
 
 ### Features
@@ -1075,6 +1112,7 @@ There are no new features introduced.
 
 - Adds `poetry` as the main project tool
 - Adds `shpinx` as a documentation tool
+
 
 ## Version 0.0.1